Refactor and optimize create class form
